10/11/2010 update 1
Moved the process of updating a list of unique individuals and updating
interaction network matrices to the class FSTcatalog. Created a class called
MMatrix (i.e. mutable matrix) that inherits from TNT's Array2D<double> in
order to enhance the functionality of Array2D to enable row and column
addition and deletion. This may eventually change to Array2D<int> or
Array2D<bool> as all entries in intxnNet are 0s and 1s.